+++ client/components/canvas/index.tsx
@@ -1,7 +1,7 @@
 /** @jsx jsx */
 import { Button, Card, Elevation } from '@blueprintjs/core';
 import { jsx } from '@emotion/core';
-import React, { Component, createRef } from 'react';
+import React, { Component, createRef, RefObject } from 'react';
 import { canvasStyle } from './style';
 
 import { Layer, Stage } from 'react-konva';
@@ -11,6 +11,7 @@ import Text from './Text';
 import TransformerComponent from './transformer';
 
 interface IAppProps {
+  onRef: RefObject<any>;
   image: HTMLImageElement | null;
   onImageCreated: (arg1: Blob) => void;
   textBlocks: { [s: string]: ITextBlocksConfigPanelState };
@@ -108,35 +109,6 @@ class Canvas extends Component<IAppProps, IAppState> {
     });
   };
 
-  dataURItoBlob = dataURI => {
-    let byteString;
-    let mimestring;
-
-    if (dataURI.split(',')[0].indexOf('base64') !== -1) {
-      byteString = atob(dataURI.split(',')[1]);
-    } else {
-      byteString = decodeURI(dataURI.split(',')[1]);
-    }
-    mimestring = dataURI
-      .split(',')[0]
-      .split(':')[1]
-      .split(';')[0];
-    const content = new Array();
-    for (let i = 0; i < byteString.length; i++) {
-      content[i] = byteString.charCodeAt(i);
-    }
-    return new Blob([new Uint8Array(content)], { type: mimestring });
-  };
-
-  onExportImageClicked = () => {
-    const imgB64 = this.stageRef.current!.toDataURL({ pixelRatio: 3 });
-    const blob = this.dataURItoBlob(imgB64);
-    this.props.onImageCreated(blob);
-    // console.log(blob);
-    // let fd = new FormData(document.forms[0]);
-    // fd.append("canvasImage", blob);
-  };
-
   render() {
     const {
       backgroundImage,
@@ -144,12 +116,12 @@ class Canvas extends Component<IAppProps, IAppState> {
       canvasWidth,
       selectedShapeName
     } = this.state;
-    const { textBlocks } = this.props;
+    const { textBlocks, onRef } = this.props;
     return (
       <Card elevation={Elevation.ONE} css={canvasStyle}>
         {process.browser && (
           <Stage
-            ref={this.stageRef}
+            ref={onRef}
             width={canvasWidth}
             height={canvasHeight}
             onMouseDown={this.handleStageMouseDown}
@@ -172,7 +144,6 @@ class Canvas extends Component<IAppProps, IAppState> {
             </Layer>
           </Stage>
         )}
-        <Button onClick={this.onExportImageClicked}> Export Image </Button>
       </Card>
     );
   }
